e2e: idempotently provision Nomad, Consul, Vault via TF
Run a target-specific install-nomad script rather than run commands over ssh
which need to be customized for each distro. Use Terraform's detection of
variable value changes so that we can re-run the provisioning without having
to re-install Nomad on those specific hosts that need it changed.

Use configuration profile directories as the source of configs. Detect changes
in the file contents so that we only upload configuration files that have
changed between Terraform runs.